+++ MaDKitLanEdition/src/test/java/com/distrimind/madkit/kernel/network/NetworkGlobalBandwidthLimitTest.java
@@ -8,13 +8,13 @@ public class NetworkGlobalBandwidthLimitTest {
 
     @Test
     public void testUploadLimit() throws UnknownHostException {
-        BigDataTransferSpeed test=new BigDataTransferSpeed(Integer.MAX_VALUE, 40000000);
+        BigDataTransferSpeed test=new BigDataTransferSpeed(Long.MAX_VALUE, 40000000);
         test.bigDataTransfer();
 
     }
     @Test
     public void testDownloadLimit() throws UnknownHostException {
-        BigDataTransferSpeed test=new BigDataTransferSpeed(40000000, Integer.MAX_VALUE);
+        BigDataTransferSpeed test=new BigDataTransferSpeed(40000000, Long.MAX_VALUE);
         test.bigDataTransfer();
     }
 }
